test(transform): separate the rejection pins from the parity pins
Review fallout, no change to any snapshot.

`class_name_edge_cases` opened by promising byte parity with
`@stylexjs/babel-plugin` and then closed with five cases that reject values
upstream accepts. Two contracts in one file, and the second read as a violation
of the first. The rejections move to `class_name_rejections`, whose own doc says
plainly that upstream accepts every value in it.

The edge helper carried a `customize` hook no test used and switched font-size
conversion on, which the harness leaves off. No value in the file is a font
size, so the snapshots matched their measurements either way — by luck rather
than by construction. Both are gone; the helper now sets only what the
measurement was taken under.

The nested value that trips the depth guard was one 448-character line. It is
now assembled from wrapped literals, and the doc states the two shapes that
still run long on purpose: a quoted rule wrapped is a quoted rule falsified, and
a value literal broken up is no longer the text the harness measured.

Both file headers cited a command that did not produce the report they were read
from. They now name the corpus set and the build step that has to precede it.
