+++ crates/stylex-transform/tests/transform_value_normalization_test/mod.rs
@@ -1,3 +1,4 @@
 mod class_name_edge_cases;
+mod class_name_rejections;
 mod css_value_normalization;
 mod transform_font_size_with;
